requirements.txt
fastapi
uvicorn
xgboost
pandas
scikit-learn
tldextract
python-whois
requests
